backend/pyproject.toml
[tool.poetry]
name = "01_sprint_01"
version = "0.1.0"
description = ""
authors = ["captainKoala <kostjan-pirat@rambler.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
flake8 = "4.0"
django-split-settings = "^1.1.0"
psycopg2-binary = "^2.9.3"
django-debug-toolbar = "^3.4.0"
pytest = "^7.1.2"
pydantic = "^1.9.1"
python-dotenv = "^0.20.0"
uWSGI = "^2.0.20"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
